% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dss_to_netcdf}
\alias{dss_to_netcdf}
\title{Convert dss to netcdf}
\usage{
dss_to_netcdf(dss, nc, variable_parts = "B")
}
\arguments{
\item{dss}{a dss file handle, from opendss}

\item{nc}{a nc file handle, from dss_to_ncdf_init}
}
\value{
Outputs a NetCDF file
}
\description{
This function does the bulk of the work to convert dss data to netcdf,
but not all of it. In particular, you will need to call \code{\link{dss_to_ncdf_init}}
first to set up the netcdf file. The way dss and netcdf handle time data
is slightly different and so some data will not convert nicely. See the
description of \code{\link{dss_to_ncdf_init}} for more information
}
\note{
NOTE
}
\author{
Cameron Bracken
}
\seealso{
\code{\link[dssrip]{opendss}}, \code{\link{dss_to_ncdf_init}}
}

