% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dss_var_to_ncdf}
\alias{dss_var_to_ncdf}
\title{Writes a dss variable to an existing netcdf file}
\usage{
dss_var_to_ncdf(v, nc, nc_datetimes = NULL)
}
\arguments{
\item{v}{a variable object returned by \code{\link{read_dss_variable}}}

\item{nc}{A RNetCDF file handle set up by \code{\link{dss_to_ncdf_init}}}

\item{nc_datetimes}{the datetimes from the time variable in the netcdf file
                    this is for efficiency so the times do not have to be
                    read and converted for each new variable written.}
}
\value{
TRUE if the write was successful, false otherwise
}
\description{
Uses the output of read_dss_variable, creating a new variable in an _existing_
netcdf file
}
\author{
Cameron Bracken
}

