% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_dss_list}
\alias{read_dss_list}
\title{dss_to_list get data from a dss file as a list of lists}
\usage{
read_dss_list(dss, variable_parts = "B")
}
\arguments{
\item{dss}{a dss file handle, from opendss}
}
\value{
list of lists containing dss data
}
\description{
Read all variable from a dss file into R
}
\details{
Read all data from a dss file into a list, one variable per element.
Be aware this will probably take a long time.
}
\note{
NOTE
}
\author{
Cameron Bracken
}
\seealso{
\code{\link{nchar}} which this function wraps
}

