% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_dss_variable}
\alias{read_dss_variable}
\title{Read a single dss variable out of a dss file}
\usage{
read_dss_variable(variable, dss, parts = NULL, variable_parts = "B")
}
\arguments{
\item{variable}{dss variable defined by the combination of one or more path parts}

\item{dss}{a dss file handle from opendss}

\item{parts}{a dss path parts data.table returned from separate_path_parts, if NULL,
it will be generated from the dss file (can take time for big dss files)}

\item{variable_parts}{the letters corresponding to the path parts to use as a unique
                      identifier for a variable, default "B"}
}
\value{
list containing dss data
}
\description{
This function will read a single variable out of a DSS file,
it will read the list of variables from the dss file and parse it
to find the path group unless you specefy a parts data.table, from
separatePathParts, this is because for big DSS files the searching
for the path group can take a few seconds.
}
\details{
A path group is a collection of DSS paths that define a single variable
in DSS. For some strange reason DSS does not want to put all its data in
a single path.

A 'variable' is a unique name refering to a collection of paths, it can be
defined by a dingle dss path part (default is "B") or two or more path parts
which will be separated by an "_"
}
\note{
NOTE
}
\author{
Cameron Bracken
}

