% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_dss_variable_by_path_group}
\alias{read_dss_variable_by_path_group}
\title{Read a variable from a dss file as defined by a collection of paths}
\usage{
read_dss_variable_by_path_group(variable, path_group, dss)
}
\value{
a list of data from dss
}
\description{
Long Description
}
\note{
NOTE
}
\author{
Cameron Bracken
}

