% Generated by roxygen2 (4.0.1): do not edit by hand
\name{separate_path_parts}
\alias{separate_path_parts}
\title{Break up dss variable paths for filtering}
\usage{
separate_path_parts(paths, variable_parts = "B")
}
\value{
a data.table with with path parts
}
\description{
Long Description
}
\note{
NOTE
}
\author{
Cameron Bracken
}

